refactor(free-search): restructure fetch logic, types, and file organization - moved fetchTours function into useToursStore as internalFetchTours - extracted all free-search–specific types/interfaces into free-search.types.ts - renamed API and utils files to free-search.api.ts and free-search.utils.ts for clarity and consistency
