+++ frontend-platform/src/stores/useTourStore.ts
@@ -1,40 +1,9 @@
 import { create } from "zustand";
-import { fetchTours } from "@/utils/mordern-search.utils";
-import { paginationType, Tour, TourFiltersType, SortToursType, TourSearchResponse, ToursPricedByType } from "@/types/types";
-import { ResultsStateType } from "@/types/free-search.types";
+import { ToursStateType } from "@/types/free-search.types";
 import { DEFAULT_PAGINATION } from "@/utils/constants.utils";
+import { getToursByCountryId, getToursByParkId } from "@/lib/api/free-search.api";
 
-
-
-export interface ToursState {
-    tours: Tour[];
-    pagination: paginationType;
-
-    resultsState: ResultsStateType;
-    isLoadingMore: boolean;
-
-    setPagination: (pagination: paginationType) => void;
-    setResultsState: (value: ResultsStateType) => void;
-    resetPagination: () => void;
-
-    loadTours: (
-        idParam: number,
-        typeParam: string,
-        filters: TourFiltersType,
-        sortBy: SortToursType,
-        pricedBy: ToursPricedByType
-    ) => Promise<void | TourSearchResponse>;
-
-    loadMoreTours: (
-        idParam: number,
-        typeParam: string,
-        filters: TourFiltersType,
-        sortBy: SortToursType,
-        pricedBy: ToursPricedByType
-    ) => Promise<void>;
-}
-
-export const useToursStore = create<ToursState>((set, get) => ({
+export const useToursStore = create<ToursStateType>((set, get) => ({
     tours: [],
     pagination: DEFAULT_PAGINATION,
 
@@ -45,37 +14,92 @@ export const useToursStore = create<ToursState>((set, get) => ({
     setResultsState: (value) => set({ resultsState: value }),
     resetPagination: () => set({ pagination: DEFAULT_PAGINATION }),
 
-    // Load tours with given filters
+    // Internal Fetch Tours Method
+    internalFetchTours: async (
+        id,
+        type,
+        paginationMeta,
+        setTourResults,
+        setPaginationMeta,
+        isLoadMore,
+        filters,
+        sortBy,
+        pricedBy
+    ) => {
+        try {
+            const pageToFetch = isLoadMore ? paginationMeta.page + 1 : paginationMeta.page;
+            const updatedPaginationMeta = { ...paginationMeta, page: pageToFetch };
+
+            let fetchedTours;
+
+            if (type === "park") {
+                fetchedTours = await getToursByParkId(
+                    id,
+                    updatedPaginationMeta,
+                    filters,
+                    sortBy,
+                    pricedBy
+                );
+            } else if (type === "country") {
+                fetchedTours = await getToursByCountryId(
+                    id,
+                    updatedPaginationMeta,
+                    filters,
+                    sortBy,
+                    pricedBy
+                );
+            } else {
+                throw new Error(`Unknown type: ${type}`);
+            }
+
+            // Append or replace
+            setTourResults((prev) =>
+                isLoadMore ? [...prev, ...fetchedTours.tours] : fetchedTours.tours
+            );
+
+            // Update pagination meta
+            setPaginationMeta(fetchedTours.pagination);
+
+            return fetchedTours;
+        } catch (err) {
+            throw err;
+        }
+    },
+
+    // Load Tours(For initial load and filters/sort change)
     loadTours: async (idParam, typeParam, filters, sortBy, pricedBy) => {
         if (!idParam || !typeParam) return;
 
         try {
-
-            return await fetchTours(
+            return await get().internalFetchTours(
                 idParam,
                 typeParam,
                 DEFAULT_PAGINATION,
                 (updateFn) =>
                     set((state) => ({
-                        tours: typeof updateFn === "function" ? updateFn(state.tours) : updateFn,
+                        tours:
+                            typeof updateFn === "function"
+                                ? updateFn(state.tours)
+                                : updateFn,
                     })),
                 (updateFn) =>
                     set((state) => ({
                         pagination:
-                            typeof updateFn === "function" ? updateFn(state.pagination) : updateFn,
+                            typeof updateFn === "function"
+                                ? updateFn(state.pagination)
+                                : updateFn,
                     })),
                 false,
                 filters,
                 sortBy,
                 pricedBy
             );
-
         } catch (err) {
             throw err;
         }
     },
 
-    // Load more tours with given filters
+    // Load Tours (For Load More)
     loadMoreTours: async (idParam, typeParam, filters, sortBy, pricedBy) => {
         if (!idParam || !typeParam) return;
 
@@ -83,18 +107,23 @@ export const useToursStore = create<ToursState>((set, get) => ({
         set({ isLoadingMore: true });
 
         try {
-            await fetchTours(
+            await get().internalFetchTours(
                 idParam,
                 typeParam,
                 pagination,
                 (updateFn) =>
                     set((state) => ({
-                        tours: typeof updateFn === "function" ? updateFn(state.tours) : updateFn,
+                        tours:
+                            typeof updateFn === "function"
+                                ? updateFn(state.tours)
+                                : updateFn,
                     })),
                 (updateFn) =>
                     set((state) => ({
                         pagination:
-                            typeof updateFn === "function" ? updateFn(state.pagination) : updateFn,
+                            typeof updateFn === "function"
+                                ? updateFn(state.pagination)
+                                : updateFn,
                     })),
                 true,
                 filters,
