+++ frontend-platform/src/types/types.ts
@@ -1,5 +1,3 @@
-import { ResultsStateType } from "./free-search.types";
-
 // Tour Related Types
 export interface Tour {
     id: number;
@@ -78,119 +76,3 @@ export interface Operator {
     dateModified: Date | null;
     tours: Tour[]
 }
-
-// Mordern Search Related Types
-export interface ParkSearchType {
-    id: number;
-    name: string;
-    country: string;
-    keyword: string;
-    type: string;
-};
-
-export interface CountrySearchType {
-    id: number;
-    name: string;
-    type: string;
-};
-
-export interface ParksCountriesType {
-    parks: ParkSearchType[];
-    countries: CountrySearchType[];
-    popularParks: ParkSearchType[];
-    trendingSearches: SuggestionType[];
-}
-
-export type SuggestionType = ParkSearchType | CountrySearchType;
-
-export interface paginationType {
-    page: number;
-    limit: number;
-    total: number;
-    totalPages: number;
-    hasMore: boolean;
-};
-
-export interface TourSearchResponse {
-    tours: Tour[];
-    pagination: paginationType;
-}
-
-export interface TourFiltersType {
-    accommodation: string[];
-    budget: [number, number];
-    duration: [number, number];
-}
-
-export interface ToursPricedByType {
-    persons: number;
-}
-
-export type SortToursType =
-    | "default"
-    | "duration_short_long"
-    | "duration_long_short"
-    | "budget_low_high"
-    | "budget_high_low";
-
-
-export type ActiveTabType = "all" | "parks" | "experiences";
-
-export type TourHandlerDeps = {
-    idParam: number;
-    typeParam: string;
-    filters: TourFiltersType;
-    sortBy: SortToursType;
-    router: any;
-    searchParams: any;
-    resetPagination: () => void;
-    loadTours: (
-        idParam: number,
-        typeParam: string,
-        filters: TourFiltersType,
-        sortBy: SortToursType,
-    ) => Promise<void>;
-    setFilters?: (filters: TourFiltersType) => void;
-    setSortBy?: (value: SortToursType) => void;
-};
-
-export interface LoadToursByParkParams {
-    selectedPark: { id: number; type: string } | null;
-    activeTab: ActiveTabType;
-    searchParams: URLSearchParams;
-    router: any;
-    setSearchItemType: (type: string) => void;
-    setSearchItemId: (id: number) => void;
-}
-
-export interface LoadInitialToursParams {
-    idInURL: number;
-    typeInURL: string;
-    activeTab: string;
-    searchParams: URLSearchParams;
-    router: any;
-    setSearchItemType: (value: string) => void;
-    setSearchItemId: (value: number) => void;
-    tourSearchUrlHandler: {
-        setActiveTab: (args: any) => void;
-        getFiltersFromUrl: (
-            params: URLSearchParams,
-            setAppliedFilters: (filters: any) => void
-        ) => {
-            filtersFromURL: any;
-            sortingFromURL: any;
-        };
-    };
-}
-
-export type FiltersFromUrlReturnType = {
-    filtersFromURL: {
-        accommodation: string[];
-        duration: [number, number];
-        budget: [number, number];
-    };
-    sortingFromURL: SortToursType;
-    pricedByFromURL: {
-        persons: number;
-    };
-};
